Fail earlier if not set OAUTH_PROVIDER_SECRETS env variable

+++ fixtures/auth-e2e/deploy.sh
@@ -64,6 +64,11 @@ echo -n "* Deploying the test app to $SITE..."
 # `|| true` so that the script doesn't fail if the the app doesn't exist
 "$METEOR" deploy -D $SITE >> $LOG 2>&1 || true
 
+if [ -z "$OAUTH_PROVIDER_SECRETS" ]; then
+    echo "Need to set \$OAUTH_PROVIDER_SECRETS"
+    exit 1
+fi
+
 echo $OAUTH_PROVIDER_SECRETS > secrets.json
 "$METEOR" deploy --settings secrets.json $SITE >> $LOG 2>&1
 rm secrets.json
